Cycles
It has been a bit! I got busy with school but foolishly decided I wanted to rewrite a lot of this code.  And then just forgot to make them into individual commits!

Changes:
Rewrote the Lexer class a bit because I hated looking at that giant switch statement.
It now uses a chain of function calls! ... I still hate this! but i hate it less!

Rewrote lexer to also go off of mostly chars not strings. some other spots were not switched.
This is just because alot of the tokens end up being single or only two characters, so it just makes sense.

Replaced alot of the spots I passed in a string as a function parameter with some form of Span when able (they're neat!).

Implemented For and While loops! no idea about foreach loops.

Extracted the original method of parsing a function body and made it into the "Parse_Body" function. made it a lil bit cleaner when I used it for the loops.