Fixed ElseStatement not just parsing to true, Implemented single line function/if/loop bodies
Fixed an issue with the "Parse_Else_Stmt" function did not set the expression it evaluated to "true" by default.

fixed an issue where boolean values did not use the passed in values when being constructed.

Modified the Parse_Body function to allow for non '{' '}' body expressions.
'if(true) print("true!)' for example.